app/backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==2.0.3
google-cloud-storage==2.9.0
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
google-generativeai==0.3.0
pillow==10.0.0
email-validator==2.0.0
pytest==7.3.1
httpx==0.24.1
